front-end/src/site/domain/presentation/site.component.spec.ts
import { TestBed, async } from '@angular/core/testing';

import { SiteComponent } from './site.component';

describe('PastoresViewComponent', () => {
  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [
        SiteComponent
      ],
    }).compileComponents();
  }));

  it('should create the site', async(() => {
    const fixture = TestBed.createComponent(SiteComponent);
    const app = fixture.debugElement.componentInstance;
    expect(app).toBeTruthy();
  }));

  it(`should have as title 'app'`, async(() => {
    const fixture = TestBed.createComponent(SiteComponent);
    const app = fixture.debugElement.componentInstance;
    expect(app.title).toEqual('app');
  }));

  it('should render title in a h1 tag', async(() => {
    const fixture = TestBed.createComponent(SiteComponent);
    fixture.detectChanges();
    const compiled = fixture.debugElement.nativeElement;
    expect(compiled.querySelector('h1').textContent).toContain('Welcome to site!');
  }));
});
